Develop

+++ HealthTourist.Common/Constants/Common/FileExtensionConstants.cs
@@ -2,17 +2,17 @@ namespace HealthTourist.Common.Constants.Common;
 
 public abstract class FileExtensionConstants
 {
-    public const int WordDocExtensionNo = 1;
-    public const int WordDocxExtensionNo = 2;
-    public const int WordRtfExtensionNo = 3;
-    public const int ExcelXlsExtensionNo = 4;
-    public const int ExcelXlsxExtensionNo = 5;
-    public const int PowerPointPptExtensionNo = 6;
-    public const int PowerPointPptxExtensionNo = 7;
-    public const int FilePdfExtensionNo = 8;
-    public const int FileTxtExtensionNo = 9;
-    public const int FileCsvExtensionNo = 10;
-    public const int ImageJpgExtensionNo = 11;
-    public const int ImageJpegExtensionNo = 12;
-    public const int ImagePngExtensionNo = 13;
+    public const byte WordDocExtensionNo = 1;
+    public const byte WordDocxExtensionNo = 2;
+    public const byte WordRtfExtensionNo = 3;
+    public const byte ExcelXlsExtensionNo = 4;
+    public const byte ExcelXlsxExtensionNo = 5;
+    public const byte PowerPointPptExtensionNo = 6;
+    public const byte PowerPointPptxExtensionNo = 7;
+    public const byte FilePdfExtensionNo = 8;
+    public const byte FileTxtExtensionNo = 9;
+    public const byte FileCsvExtensionNo = 10;
+    public const byte ImageJpgExtensionNo = 11;
+    public const byte ImageJpegExtensionNo = 12;
+    public const byte ImagePngExtensionNo = 13;
 }
